requirements.txt
django==4.0.6djangorestframework==3.13.1django-filter==22.1